CTJS: Load modules earlier, allowing for initial gameLoad to trigger
This reverts parts of  64525a3ededa95a414f99e78b91d87fabed35e06 and f19bec16624852ea4768f054160429948b8d415b
