+++ src/main/kotlin/com/chattriggers/ctjs/CTJS.kt
@@ -10,20 +10,14 @@ import com.chattriggers.ctjs.api.message.ChatLib
 import com.chattriggers.ctjs.api.render.Image
 import com.chattriggers.ctjs.api.triggers.TriggerType
 import com.chattriggers.ctjs.api.world.World
+import com.chattriggers.ctjs.engine.Console
 import com.chattriggers.ctjs.engine.Register
 import com.chattriggers.ctjs.internal.commands.StaticCommand
-import com.chattriggers.ctjs.engine.Console
-import com.chattriggers.ctjs.internal.console.ConsoleHostProcess
-import com.chattriggers.ctjs.internal.engine.CTEvents
 import com.chattriggers.ctjs.internal.engine.module.ModuleManager
 import com.chattriggers.ctjs.internal.utils.Initializer
 import kotlinx.serialization.json.Json
 import net.fabricmc.api.ClientModInitializer
-import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper
 import net.fabricmc.loader.api.FabricLoader
-import net.minecraft.client.option.KeyBinding
-import net.minecraft.client.util.InputUtil
-import org.lwjgl.glfw.GLFW
 import java.io.File
 import java.net.URL
 import java.net.URLConnection
@@ -33,39 +27,19 @@ import kotlin.concurrent.thread
 
 class CTJS : ClientModInitializer {
     override fun onInitializeClient() {
-        // Need to initialize the keybind here, as putting it inside the MinecraftClient::send
-        // runs it too late
-        val keybind = KeyBindingHelper.registerKeyBinding(
-            KeyBinding(
-                "chattriggers.key.binding.console",
-                InputUtil.Type.KEYSYM,
-                GLFW.GLFW_KEY_GRAVE_ACCENT,
-                "chattriggers.key.category",
-            )
-        )
+        Client.referenceSystemTime = System.nanoTime()
+        Initializer.initializers.forEach(Initializer::init)
 
-        CTEvents.RENDER_GAME.register {
-            if (keybind.wasPressed())
-                ConsoleHostProcess.show()
+        thread {
+            reportHashedUUID()
         }
 
-        Client.getMinecraft().send {
-            Client.referenceSystemTime = System.nanoTime()
+        Config.loadData()
 
-            Initializer.initializers.forEach(Initializer::init)
-
-            thread {
-                ModuleManager.entryPass()
-                reportHashedUUID()
-            }
-
-            Config.loadData()
-
-            Runtime.getRuntime().addShutdownHook(Thread {
-                TriggerType.GAME_UNLOAD.triggerAll()
-                Console.close()
-            })
-        }
+        Runtime.getRuntime().addShutdownHook(Thread {
+            TriggerType.GAME_UNLOAD.triggerAll()
+            Console.close()
+        })
     }
 
     private fun reportHashedUUID() {
