+++ src/main/kotlin/com/chattriggers/ctjs/internal/console/ConsoleHostProcess.kt
@@ -5,10 +5,16 @@ import com.chattriggers.ctjs.api.Config
 import com.chattriggers.ctjs.api.client.Client
 import com.chattriggers.ctjs.engine.LogType
 import com.chattriggers.ctjs.engine.printTraceToConsole
+import com.chattriggers.ctjs.internal.engine.CTEvents
 import com.chattriggers.ctjs.internal.engine.JSLoader
+import com.chattriggers.ctjs.internal.utils.Initializer
 import gg.essential.universal.UDesktop
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
+import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper
+import net.minecraft.client.option.KeyBinding
+import net.minecraft.client.util.InputUtil
+import org.lwjgl.glfw.GLFW
 import java.awt.Color
 import java.io.BufferedReader
 import java.io.File
@@ -37,7 +43,7 @@ import kotlin.io.path.Path
  * Each console gets its own Host/Client pair running on their own port, so there will be
  * `<number of loaders> + 1` sockets (the extra 1 is for the generic console).
  */
-object ConsoleHostProcess {
+object ConsoleHostProcess : Initializer {
     private var PORT = 9002
     private var running = true
     private lateinit var socketOut: PrintWriter
@@ -51,6 +57,22 @@ object ConsoleHostProcess {
 
     init {
         thread { hostMain() }
+    }
+
+    override fun init() {
+        val keybind = KeyBindingHelper.registerKeyBinding(
+            KeyBinding(
+                "chattriggers.key.binding.console",
+                InputUtil.Type.KEYSYM,
+                GLFW.GLFW_KEY_GRAVE_ACCENT,
+                "chattriggers.key.category",
+            )
+        )
+
+        CTEvents.RENDER_GAME.register {
+            if (keybind.wasPressed())
+                show()
+        }
     }
 
     private fun hostMain() {
