+++ src/main/kotlin/com/chattriggers/ctjs/internal/utils/Initializer.kt
@@ -5,6 +5,7 @@ import com.chattriggers.ctjs.api.client.KeyBind
 import com.chattriggers.ctjs.api.commands.DynamicCommands
 import com.chattriggers.ctjs.internal.commands.CTCommand
 import com.chattriggers.ctjs.internal.commands.StaticCommand
+import com.chattriggers.ctjs.internal.console.ConsoleHostProcess
 import com.chattriggers.ctjs.internal.engine.module.ModuleUpdater
 import com.chattriggers.ctjs.internal.listeners.ClientListener
 import com.chattriggers.ctjs.internal.listeners.MouseListener
@@ -16,6 +17,7 @@ internal interface Initializer {
     companion object {
         internal val initializers = listOf(
             ClientListener,
+            ConsoleHostProcess,
             CPS,
             CTCommand,
             DynamicCommands,
